requirements.txt
langchain
openai
tiktoken
azure-search-documents
python-dotenv
langchain-community
